adding unit tests and flake8 and black cleanup
Unit tests added for the backend. Previous unit tests corrected to run.
Starting a new game also returns the unfilled template as last item.
Flake8 and Black done to make the files happy. Docstrings should be
added to the unit test files, but have not been. Test strap for main
prettied up to have labels and lay out the data in a readable way.